create selector to isolate getting user eth reserve amount. update cash out modal to only convert up to eth reserve not all the eth in the user's signing wallet
