+++ packages/augur-ui/src/modules/auth/containers/connect-dropdown.ts
@@ -5,16 +5,12 @@ import { updateModal } from 'modules/modal/actions/update-modal';
 import { MODAL_GAS_PRICE, GAS_SPEED_LABELS, GAS_TIME_LEFT_LABELS, MODAL_ADD_FUNDS, MODAL_UNIVERSE_SELECTOR } from 'modules/common/constants';
 import { NULL_ADDRESS } from '@augurproject/sdk/src/state/getter/types';
 import { FormattedNumber } from 'modules/types';
-import { DESIRED_SIGNER_ETH_BALANCE } from 'contract-dependencies-gsn/src/ContractDependenciesGSN';
-import { formatAttoEth, formatEther } from 'utils/format-number';
 import { AppState } from 'appStore';
-import { createBigNumber } from 'utils/create-big-number';
+import { getEthReserve } from 'modules/auth/selectors/get-eth-reserve';
 
 const mapStateToProps = (state: AppState) => {
   const { fast, average, safeLow, userDefinedGasPrice } = state.gasPriceInfo;
-  const { loginAccount } = state;
-  const { balances } = loginAccount;
-  const ethNonSafeBN = createBigNumber(balances.ethNonSafe);
+
   const userDefined = userDefinedGasPrice || average || 0;
   let gasPriceSpeed = GAS_SPEED_LABELS.STANDARD;
   let gasPriceTime = GAS_TIME_LEFT_LABELS.STANDARD
@@ -29,12 +25,7 @@ const mapStateToProps = (state: AppState) => {
     gasPriceSpeed = GAS_SPEED_LABELS.SLOW;
   }
 
-  let desiredSignerEthBalance = createBigNumber(formatAttoEth(Number(DESIRED_SIGNER_ETH_BALANCE)).value);
-  if (ethNonSafeBN.lt(desiredSignerEthBalance)) desiredSignerEthBalance = ethNonSafeBN;
-  const reserveEthAmount: FormattedNumber = formatEther(desiredSignerEthBalance, {
-    zeroStyled: false,
-    decimalsRounded: 4,
-  });
+  const reserveEthAmount: FormattedNumber = getEthReserve(state);
 
   return {
     universeOutcomeName: state.universe.outcomeName ? state.universe.outcomeName : null,
