+++ packages/augur-ui/src/modules/modal/containers/modal-cash-out.ts
@@ -6,20 +6,23 @@ import { closeModal } from 'modules/modal/actions/close-modal';
 import { ThunkDispatch } from 'redux-thunk';
 import { Action } from 'redux';
 import { withdrawAllFunds, withdrawAllFundsEstimateGas } from 'modules/contracts/actions/contractCalls';
+import { FormattedNumber } from 'modules/types';
+import { getEthReserve } from 'modules/auth/selectors/get-eth-reserve';
 
 const mapStateToProps = (state: AppState) => {
   const { loginAccount, appStatus, modal } = state;
 
-  const { address, meta } = loginAccount;
-  const signerAccount = meta.signer._address;
+  const { address } = loginAccount;
+
+  const ethReserveAmount: FormattedNumber = getEthReserve(state);
 
   return {
     account: address,
-    signerAccount,
     modal,
     GsnEnabled: appStatus.gsnEnabled,
     ethToDaiRate: appStatus.ethToDaiRate,
     gasPrice: state.gasPriceInfo.userDefinedGasPrice || state.gasPriceInfo.average,
+    ethReserveAmount
   }
 };
 
@@ -33,7 +36,7 @@ const mergeProps = (sP: any, dP: any, oP: any) => ({
   GsnEnabled: sP.GsnEnabled,
   balances: sP.balances,
   account: sP.account,
-  signerAccount: sP.signerAccount,
+  ethReserveAmount: sP.ethReserveAmount,
   ethToDaiRate: sP.ethToDaiRate,
   gasPrice: sP.gasPrice,
   closeAction: () => dP.closeModal(),
